requirements.txt
baidu-aip
dlib
face-alignment
gradio
matplotlib
numpy
opencv-python
pandas
pillow
pydantic
pyvista
scikit-image
scikit-learn
scipy
